SmrPlanet: fix creditCurrentAttackersForKill
The `player_attacks_planet.level` field is a float, not an int. Since
more strict validation was added in 1644e26e1b, this was causing an
error, e.g.:

> Failed to convert '41.6667' to int

Prior to 1644e26e1b, we were inadvertently rounding the planet levels
added to the Combat:Planet:Levels HOF since d6715e7ebe (2011).

+++ src/lib/Default/SmrPlanet.php
@@ -1317,7 +1317,7 @@ class SmrPlanet {
 		$dbResult = $this->db->read('SELECT account_id,level FROM player_attacks_planet WHERE ' . $this->SQL . ' AND time > ' . $this->db->escapeNumber(Smr\Epoch::time() - self::TIME_TO_CREDIT_BUST));
 		foreach ($dbResult->records() as $dbRecord) {
 			$currPlayer = SmrPlayer::getPlayer($dbRecord->getInt('account_id'), $this->getGameID());
-			$currPlayer->increaseHOF($dbRecord->getInt('level'), ['Combat', 'Planet', 'Levels'], HOF_PUBLIC);
+			$currPlayer->increaseHOF($dbRecord->getFloat('level'), ['Combat', 'Planet', 'Levels'], HOF_PUBLIC);
 			$currPlayer->increaseHOF(1, ['Combat', 'Planet', 'Completed'], HOF_PUBLIC);
 		}
 		$this->db->write('DELETE FROM player_attacks_planet WHERE ' . $this->SQL);
